Fix book removal issue
Previous code caused the inability to remove more than 1 book from the
list (or any book) in any situation. This was due to addEventListener
being incompatible with dynamic HTML elements, as 'click' was not
registering for new html elements.

Solution is to target a container instead of each dynamically added
'remove' button in a NodeList.